requirements.txt
groq>=0.4.0
upstash-vector>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
